rename and style fixes
