+++ src/main/java/com/artisan/solutions/gallery/service/ArtistService.java
@@ -2,6 +2,6 @@ package com.artisan.solutions.gallery.service;
 
 import com.artisan.solutions.gallery.persistence.model.Artist;
 
-public interface ArtistService extends Crud<Artist> {
+public interface ArtistService extends CrudService<Artist> {
 
 }
