+++ src/main/java/com/artisan/solutions/gallery/service/WorkService.java
@@ -2,6 +2,6 @@ package com.artisan.solutions.gallery.service;
 
 import com.artisan.solutions.gallery.persistence.model.Work;
 
-public interface WorkService extends Crud<Work> {
+public interface WorkService extends CrudService<Work> {
 
 }
