+++ src/main/java/com/artisan/solutions/gallery/spring/PersistenceJpaConfig.java
@@ -31,8 +31,6 @@ public class PersistenceJpaConfig {
         super();
     }
 
-    // beans
-
     @Bean
     public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
         final LocalContainerEntityManagerFactoryBean em = new LocalContainerEntityManagerFactoryBean();
@@ -66,15 +64,10 @@ public class PersistenceJpaConfig {
         return new PersistenceExceptionTranslationPostProcessor();
     }
 
-    //
-
     final Properties additionalProperties() {
         final Properties hibernateProperties = new Properties();
         hibernateProperties.setProperty("hibernate.hbm2ddl.auto", env.getProperty("hibernate.hbm2ddl.auto", "create-drop"));
         hibernateProperties.setProperty("hibernate.dialect", env.getProperty("hibernate.dialect"));
-
-        // setProperty("hibernate.hbm2ddl.auto", hibernateHbm2ddlAuto);
-        // setProperty("hibernate.ejb.naming_strategy", org.hibernate.cfg.ImprovedNamingStrategy.class.getName());
         return hibernateProperties;
     }
